backend/requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
Django==4.2.11
django-extensions==3.2.3
django-ninja==1.1.0
pydantic==2.7.1
pydantic_core==2.18.2
sqlparse==0.5.0
typing_extensions==4.11.0
langchain==0.0.353
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.18.1
openai==0.27.6
faiss-cpu==1.7.4
altair==4
tiktoken==0.4.0
